Move Impl, remove request_index
